Depend on Apache httpclient 4 plugin instead of jar file
https://plugins.jenkins.io/apache-httpcomponents-client-4-api/ avoids
the plugin bundling an outdated version of the Apache httpclient library.
It replaces the bundling of the jar file with a dependency on the Apache
HttpComponents Client 4.x API plugin.  That plugin is installed on over
96% of all Jenkins installations.

By depending on the plugin instead of the jar file, the plugin bill of
materials will manage the dependency version and the plugin will use
the most recent version that is installed on the Jenkins controller.

This is a common technique used by many Jenkins plugins.  Refer to
https://plugins.jenkins.io/apache-httpcomponents-client-4-api/dependencies/
for the list of plugins that depend on the Apache httpclient plugin.