Use development version numbers
Version 1.15.0 and 1.15.1 of the plugin have been released to the
Jenkins update center, even though there is no tag in the repository
for those releases.

Avoid user confusion by setting the version number to 1.15.2-SNAPSHOT.

When the maintainer performs `mvn release:prepare release:perform`,
Apache Maven will prompt for a new version number and will suggest 1.15.2.
The maintainer can press enter to accept the recommendation and Apache
Maven will create two commits, one that sets the release to 1.15.2 and
then a second that sets it to 1.15.3-SNAPSHOT.